create database file
