+++ src/EMSx.jl
@@ -6,7 +6,10 @@ using Dates, CSV, DataFrames
 using JLD
 using HTTP
 
+import DataDeps: unpack
+
 include("struct.jl")
+include("database_interface.jl")
 include("function.jl")
 include("simulate.jl")
